requirements.txt
tensorflow
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
pillow
streamlit
chromadb
kagglehub
fpdf
